+++ lido/methods/__init__.py
@@ -1,6 +1,7 @@
-from lido.methods.keys import find_keys_duplicates, validate_keys, validate_key
+from lido.methods.keys import find_duplicated_keys, validate_keys, validate_key
 from lido.methods.operators import (
-    get_operators_count,
+    get_operators_indexes,
     get_operators_data,
     get_operators_keys,
 )
+from lido.methods.stats import get_status