+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This changelog format is based on [Keep a Changelog](https://keepachangelog.com/
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/lidofinance/lido-python-sdk)
+### Added
+- Lido public class [#0020](https://github.com/lidofinance/lido-python-sdk/pull/20)
 
 ## [0.3.0](https://github.com/lidofinance/lido-python-sdk) - 2021-10-26
 ### Fixed